web_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.196
opencv-python==4.8.1.78
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.21.0
Pillow>=8.0.0
pydantic>=2.0.0
python-multipart>=0.0.6 